[AL-5942] Moving tests around to the designated folders (#1203)
