+++ tests/integration/export_v2/test_export_data_rows.py
@@ -14,11 +14,24 @@ def test_export_data_rows(client, data_row, wait_for_data_row_processing):
         "project_details": True,
         "data_row_details": True
     }
+
     task = DataRow.export_v2(client=client, data_rows=[data_row])
     task.wait_till_done()
     assert task.status == "COMPLETE"
     assert task.errors is None
     assert len(task.result) == 1
-    assert task.result[0]["data_row"]["id"] == data_row.uid
-    assert task.result[0]["data_row"]["external_id"] == data_row.external_id
-    assert task.result[0]["data_row"]["row_data"] == data_row.row_data
+    assert task.result[0]['data_row']['id'] == data_row.uid
+
+    task = DataRow.export_v2(client=client, data_rows=[data_row.uid])
+    task.wait_till_done()
+    assert task.status == "COMPLETE"
+    assert task.errors is None
+    assert len(task.result) == 1
+    assert task.result[0]['data_row']['id'] == data_row.uid
+
+    task = DataRow.export_v2(client=client, global_keys=[data_row.global_key])
+    task.wait_till_done()
+    assert task.status == "COMPLETE"
+    assert task.errors is None
+    assert len(task.result) == 1
+    assert task.result[0]['data_row']['id'] == data_row.uid
