+++ tests/integration/test_data_rows.py
@@ -9,7 +9,6 @@ import requests
 
 from labelbox import DataRow
 from labelbox.exceptions import MalformedQueryException
-from labelbox.schema.export_filters import DatarowExportFilters
 from labelbox.schema.task import Task
 from labelbox.schema.data_row_metadata import DataRowMetadataField, DataRowMetadataKind
 import labelbox.exceptions
@@ -1037,30 +1036,3 @@ def test_create_data_row_with_media_type(dataset, image_url, is_adv_enabled):
             exc.value)
 
     dataset.create_data_row(row_data=image_url, media_type="IMAGE")
-
-
-def test_export_data_rows(client, data_row, wait_for_data_row_processing):
-    # Ensure created data rows are indexed
-    data_row = wait_for_data_row_processing(client, data_row)
-    time.sleep(7)  # temp fix for ES indexing delay
-
-    task = DataRow.export_v2(client=client, data_rows=[data_row])
-    task.wait_till_done()
-    assert task.status == "COMPLETE"
-    assert task.errors is None
-    assert len(task.result) == 1
-    assert task.result[0]['data_row']['id'] == data_row.uid
-
-    task = DataRow.export_v2(client=client, data_rows=[data_row.uid])
-    task.wait_till_done()
-    assert task.status == "COMPLETE"
-    assert task.errors is None
-    assert len(task.result) == 1
-    assert task.result[0]['data_row']['id'] == data_row.uid
-
-    task = DataRow.export_v2(client=client, global_keys=[data_row.global_key])
-    task.wait_till_done()
-    assert task.status == "COMPLETE"
-    assert task.errors is None
-    assert len(task.result) == 1
-    assert task.result[0]['data_row']['id'] == data_row.uid
