+++ tests/integration/test_dataset.py
@@ -144,59 +144,6 @@ def test_bulk_conversation(dataset, sample_bulk_conversation: list) -> None:
     assert len(list(dataset.data_rows())) == len(sample_bulk_conversation)
 
 
-def test_data_row_export(dataset, image_url):
-    n_data_rows = 2
-    ids = set()
-    for _ in range(n_data_rows):
-        ids.add(dataset.create_data_row(row_data=image_url))
-    result = list(dataset.export_data_rows())
-    assert len(result) == n_data_rows
-    assert set(result) == ids
-
-
-@pytest.mark.parametrize('data_rows', [3], indirect=True)
-def test_dataset_export_v2(export_v2_test_helpers, dataset, data_rows):
-    data_row_ids = [dr.uid for dr in data_rows]
-    params = {"performance_details": False, "label_details": False}
-    task_results = export_v2_test_helpers.run_dataset_export_v2_task(
-        dataset, params=params)
-    assert len(task_results) == 3
-    assert set([dr['data_row']['id'] for dr in task_results
-               ]) == set(data_row_ids)
-
-    # testing with a datarow ids filter
-    datarow_filter_size = 2
-    data_row_ids = [dr.uid for dr in data_rows]
-
-    params = {"performance_details": False, "label_details": False}
-    filters = {"data_row_ids": data_row_ids[:datarow_filter_size]}
-
-    task_results = export_v2_test_helpers.run_dataset_export_v2_task(
-        dataset, filters=filters, params=params)
-
-    # only 2 datarows should be exported
-    assert len(task_results) == datarow_filter_size
-    # only filtered datarows should be exported
-    assert set([dr['data_row']['id'] for dr in task_results
-               ]) == set(data_row_ids[:datarow_filter_size])
-
-    # testing with a global key and a datarow id filter
-    datarow_filter_size = 2
-    global_keys = [dr.global_key for dr in data_rows]
-
-    params = {"performance_details": False, "label_details": False}
-    filters = {"global_keys": global_keys[:datarow_filter_size]}
-
-    task_results = export_v2_test_helpers.run_dataset_export_v2_task(
-        dataset, filters=filters, params=params)
-
-    # only 2 datarows should be exported
-    assert len(task_results) == datarow_filter_size
-    # only filtered datarows should be exported
-    assert set([dr['data_row']['global_key'] for dr in task_results
-               ]) == set(global_keys[:datarow_filter_size])
-
-
 def test_create_descriptor_file(dataset):
     import unittest.mock as mock
     with mock.patch.object(dataset.client,
